refactor: update ui folder to components/ui
