+++ docs/src/content/docs/ui-and-theme/Forms.mdx
@@ -57,7 +57,7 @@ import React from 'react';
 import { useForm } from 'react-hook-form';
 import * as z from 'zod';
 
-import { useAuth } from '@/core';
+import { useAuth } from '@/lib';
 import { Button, ControlledInput, View } from '@/ui';
 
 const schema = z.object({
