+++ src/app/(app)/style.tsx
@@ -4,7 +4,7 @@ import { Buttons } from '@/components/buttons';
 import { Colors } from '@/components/colors';
 import { Inputs } from '@/components/inputs';
 import { Typography } from '@/components/typography';
-import { FocusAwareStatusBar, SafeAreaView, ScrollView } from '@/ui';
+import { FocusAwareStatusBar, SafeAreaView, ScrollView } from '@/components/ui';
 
 export default function Style() {
   return (
