+++ src/app/[...messing].tsx
@@ -1,6 +1,6 @@
 import { Link, Stack } from 'expo-router';
 
-import { Text, View } from '@/ui';
+import { Text, View } from '@/components/ui';
 
 export default function NotFoundScreen() {
   return (
