+++ src/app/login.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 
 import type { LoginFormProps } from '@/components/login-form';
 import { LoginForm } from '@/components/login-form';
-import { useAuth } from '@/core';
-import { FocusAwareStatusBar } from '@/ui';
+import { FocusAwareStatusBar } from '@/components/ui';
+import { useAuth } from '@/lib';
 
 export default function Login() {
   const router = useRouter();
