+++ src/app/onboarding.tsx
@@ -2,8 +2,14 @@ import { useRouter } from 'expo-router';
 import React from 'react';
 
 import { Cover } from '@/components/cover';
-import { useIsFirstTime } from '@/core/hooks';
-import { Button, FocusAwareStatusBar, SafeAreaView, Text, View } from '@/ui';
+import {
+  Button,
+  FocusAwareStatusBar,
+  SafeAreaView,
+  Text,
+  View,
+} from '@/components/ui';
+import { useIsFirstTime } from '@/lib/hooks';
 export default function Onboarding() {
   const [_, setIsFirstTime] = useIsFirstTime();
   const router = useRouter();
