+++ src/components/buttons.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Button, View } from '@/ui';
+import { Button, View } from '@/components/ui';
 
 import { Title } from './title';
