+++ src/components/card.tsx
@@ -2,7 +2,7 @@ import { Link } from 'expo-router';
 import React from 'react';
 
 import type { Post } from '@/api';
-import { Image, Pressable, Text, View } from '@/ui';
+import { Image, Pressable, Text, View } from '@/components/ui';
 
 type Props = Post;
