+++ src/components/colors.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { Text, View } from '@/ui';
-import colors from '@/ui/colors';
+import { Text, View } from '@/components/ui';
+import colors from '@/components/ui/colors';
 
 import { Title } from './title';
 type ColorName = keyof typeof colors;
