+++ src/components/settings/language-item.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 
-import { useSelectedLanguage } from '@/core';
-import { translate } from '@/core';
-import type { Language } from '@/core/i18n/resources';
-import type { OptionType } from '@/ui';
-import { Options, useModal } from '@/ui';
+import type { OptionType } from '@/components/ui';
+import { Options, useModal } from '@/components/ui';
+import { useSelectedLanguage } from '@/lib';
+import { translate } from '@/lib';
+import type { Language } from '@/lib/i18n/resources';
 
 import { Item } from './item';
 
@@ -16,7 +16,7 @@ export const LanguageItem = () => {
       setLanguage(option.value as Language);
       modal.dismiss();
     },
-    [setLanguage, modal]
+    [setLanguage, modal],
   );
 
   const langs = React.useMemo(
@@ -24,12 +24,12 @@ export const LanguageItem = () => {
       { label: translate('settings.english'), value: 'en' },
       { label: translate('settings.arabic'), value: 'ar' },
     ],
-    []
+    [],
   );
 
   const selectedLanguage = React.useMemo(
     () => langs.find((lang) => lang.value === language),
-    [language, langs]
+    [language, langs],
   );
 
   return (
