+++ src/components/settings/theme-item.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import type { ColorSchemeType } from '@/core';
-import { translate, useSelectedTheme } from '@/core';
-import type { OptionType } from '@/ui';
-import { Options, useModal } from '@/ui';
+import type { OptionType } from '@/components/ui';
+import { Options, useModal } from '@/components/ui';
+import type { ColorSchemeType } from '@/lib';
+import { translate, useSelectedTheme } from '@/lib';
 
 import { Item } from './item';
 
@@ -16,7 +16,7 @@ export const ThemeItem = () => {
       setSelectedTheme(option.value as ColorSchemeType);
       modal.dismiss();
     },
-    [setSelectedTheme, modal]
+    [setSelectedTheme, modal],
   );
 
   const themes = React.useMemo(
@@ -25,12 +25,12 @@ export const ThemeItem = () => {
       { label: `${translate('settings.theme.light')} 🌞`, value: 'light' },
       { label: `${translate('settings.theme.system')} ⚙️`, value: 'system' },
     ],
-    []
+    [],
   );
 
   const theme = React.useMemo(
     () => themes.find((t) => t.value === selectedTheme),
-    [selectedTheme, themes]
+    [selectedTheme, themes],
   );
 
   return (
