+++ src/components/typography.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Text, View } from '@/ui';
+import { Text, View } from '@/components/ui';
 
 import { Title } from './title';
