+++ src/components/ui/button.test.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { Text } from 'react-native';
 
-import { cleanup, render, screen, setup } from '@/core/test-utils';
+import { cleanup, render, screen, setup } from '@/lib/test-utils';
 
 import { Button } from './button';
