+++ src/components/ui/checkbox.tsx
@@ -8,7 +8,7 @@ import {
 } from 'react-native';
 import Svg, { Path } from 'react-native-svg';
 
-import colors from '@/ui/colors';
+import colors from '@/components/ui/colors';
 
 import { Text } from './text';
 
@@ -159,9 +159,7 @@ export const RadioIcon = ({ checked = false }: IconProps) => {
       transition={{ borderColor: { duration: 100, type: 'timing' } }}
     >
       <MotiView
-        className={`h-[10px] w-[10px] rounded-[10px] ${
-          checked && 'bg-primary-300'
-        } `}
+        className={`size-[10px] rounded-[10px] ${checked && 'bg-primary-300'} `}
         from={{ opacity: 0 }}
         animate={{ opacity: checked ? 1 : 0 }}
         transition={{ opacity: { duration: 50, type: 'timing' } }}
@@ -219,7 +217,7 @@ export const SwitchIcon = ({ checked = false }: IconProps) => {
         />
       </View>
       <MotiView
-       style={{
+        style={{
           height: THUMB_HEIGHT,
           width: THUMB_WIDTH,
           position: 'absolute',
