+++ src/components/ui/focus-aware-status-bar.tsx
