+++ src/components/ui/icons/settings.tsx
