+++ src/components/ui/icons/style.tsx
