+++ tsconfig.json
@@ -5,7 +5,7 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["./src/*"],
-      "@env": ["./src/core/env.js"]
+      "@env": ["./src/lib/env.js"]
     },
     "esModuleInterop": true,
     "checkJs": true
